genomegenie/io: don't accept VariantRecord iterator in to_arrow
`pysam.VariantFile` is not pickleable (no __reduce__ implementation),
do not accept iterators over the file records.  Instead, pass the
options, and open a new `VariantFile` object inside each call of
`to_arrow`.  This adds an overhead as the same file is opened multiple
times, and the header has to be decompressed and read every time.