io.to_arrow: parse sample wise data to write to separte file
The idea is, typically we don't want to read the sample wise data.
When we do want to, we want to read it for all samples, but for
specific regions of the genome.  So essentially it is transposed in
relation to the primary variant call columns.

FIXME: It's a complete mess :-|